backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
cloudinary==1.36.0
google-generativeai==0.3.2
python-dotenv==1.0.1

# AI/ML Stack
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
ultralytics==8.1.11
sentence-transformers==2.3.1
open-clip-torch==2.24.0
Pillow==10.2.0
opencv-python==4.9.0.80
imagehash==4.3.1
numpy==1.26.3

# Vector Database
faiss-cpu

# Background Tasks
celery==5.3.6
redis==5.0.1
firebase-admin==6.4.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
geoalchemy2==0.14.3

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
requests==2.31.0
httpx==0.25.2
loguru==0.7.2
pydantic==2.5.3
pydantic-settings==2.1.0
